Start demo config file
This commit starts the process of building the demo home-assistant
configuration. It adds a few demo devices (we can easily expand or
contract this set) and adds the scaffolding for the watson_iot
component. That can't have a hard coded configuration in the repo
because all it's values are fully dependent on how the iot platform
board is setup.