Add Watson IoT Platform details to README
This commit adds details to the README file for setting up the Watson
IoT Platform and configuring Home Assistant to use it. This borrows
heavily from the README in the IBM/guide-conveyor-simulator
(https://github.com/IBM/guide-conveyor-simulator/blob/master/README.md)
for the watson setup. These steps are the same and the README there
wrote it more eloquently than I was capable of.