Cleanup config directory
This commit cleans up some aspects of the config directory. First the
example configuration file now has the watson iot component
configuration commented out. This is because by default we can't have it
enabled without knowing the watson iot credentials. If we want people to
be able to use the workflow:
setup hass -> setup watson -> configure hass for watson -> use watson
we need to having a working config to start with.

THe other aspect here is the custom component for the watson iot usage
needs to be including in the config dir per:
https://developers.home-assistant.io/docs/en/creating_component_loading.html
This just moves the component to proper relative directory to make
automated hass deployment easier.