Add README section for running home assistant locally
This commit adds a new section to the documentation on how to run home
assistant as a demo locally. Included in this are support scripts to
automate the setup as much as possible for running a demo. These
sections will be expanded upon in the future to cover more about
customizing the demo environment. Additionally a section will be added
about running home assistant in IBM cloud's container service.