Fix setup script
This commit fixes a few issues in the setup.sh script for getting a
local home assistant demo running in a venv. There were also some
configuration issues which prevented the frontend from running so when
launching the script opens the browser it would get a 404 from the hass
server.