Use IBM Code Pattern template
This commit adjusts the readme to use the IBM code pattern template. As
part of this the code in the project is now licensed as Apache 2.0
explictly, which was missing from the repo before. (but was implied
because of Home-Assistant's license and the fact the custom module
intree has been proposed as an addition to the upstream project)