Rename custom component
With the upstream review of the watson_iot component progressing, but
moving slowly this commit takes precautions for when it eventually
merges. While the components do the same thing and essential the same we
don't want this demo to stop working for people when upstream hass
includes a watson_iot component in a release. To avoid any potential
problems this just renames the component to be watson_iot_platform
instead.