Add initial cloud based deploy support to pattern
This commit adds the initial work for doing a cloud based deploy of Home
Assistant. This is really only useful for the demo, but it will give
people the closest to a one-click solution for experimenting with
things. There are still manual setup steps required because of the
Watson IoT Platform. But this will at least enable some degree of
automation without a local environment.