Pre-install ibmiotf dependency for home assistant
Home assistant dynamically calls pip during runtime based on the
components that are enabled in configuration. This way you don't have to
install several hundred optional dependencies unless you're using that
component. However, when we were enabling the the iotf plugin post
container creation the download from pypi was being blocked. This caused
the watson_iot_platform custom component to error on launch, because it
couldn't call the ibmiotf library. This commit works around this issue
by pre-installing ibmiotf in the image build so that the sdist will be
cached in the container for when home assistant installs it during
runtime.